rename handleExitjs to handleExit.js
