+++ src/common/utils/handleExit.js
